requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
opencv-python==4.5.5.64
numpy==1.21.6
easyocr==1.4.1
onnxruntime==1.10.0
Pillow==8.2.0